share _wrap_joined_index
